+++ src/formatters/luau.rs
@@ -401,6 +401,19 @@ pub fn hang_type_info(ctx: &Context, type_info: &TypeInfo, shape: Shape) -> Type
                 shape.reset() + PIPE_LENGTH,
             )),
         },
+        TypeInfo::Intersection {
+            left,
+            ampersand,
+            right,
+        } => TypeInfo::Intersection {
+            left: Box::new(format_type_info(ctx, left, shape)),
+            ampersand: hang_type_info_binop(ctx, ampersand.to_owned(), shape, right),
+            right: Box::new(hang_type_info(
+                ctx,
+                &right.update_leading_trivia(FormatTriviaType::Replace(vec![])),
+                shape.reset() + PIPE_LENGTH,
+            )),
+        },
         other => format_type_info(ctx, other, shape),
     }
 }
@@ -535,12 +548,21 @@ pub fn format_type_assertion(
 fn should_hang_type(type_info: &TypeInfo) -> bool {
     // Only hang if its a binary type info, since it doesn't matter for unary types
     match type_info {
-        TypeInfo::Union { left, pipe, right } => {
+        TypeInfo::Union {
+            left,
+            pipe: binop,
+            right,
+        }
+        | TypeInfo::Intersection {
+            left,
+            ampersand: binop,
+            right,
+        } => {
             type_info_trailing_trivia(left)
                 .iter()
                 .any(trivia_is_comment)
                 || should_hang_type(left)
-                || contains_comments(pipe)
+                || contains_comments(binop)
                 || full_moon::node::Node::surrounding_trivia(right)
                     .0
                     .iter()