+++ src/shape.rs
@@ -1,4 +1,11 @@
 use crate::context::Context;
+#[cfg(feature = "luau")]
+use crate::formatters::{
+    trivia::{FormatTriviaType, UpdateTrivia},
+    trivia_util::trivia_is_comment,
+};
+#[cfg(feature = "luau")]
+use full_moon::node::Node;
 use std::fmt::Display;
 use std::ops::Add;
 
@@ -204,9 +211,36 @@ impl Shape {
     }
 
     /// Takes in a new node, and tests whether adding it in will force any lines over the budget.
+    /// This function attempts to ignore the impact of comments by removing them, which makes this function more expensive.
     /// NOTE: This function does not update state/return a new shape
-    pub fn test_over_budget<T: Display>(&self, item: &T) -> bool {
-        let string = format!("{}", item);
+    #[cfg(feature = "luau")]
+    pub fn test_over_budget<T: Node>(&self, item: &T) -> bool {
+        // Converts the node into a string, removing any comments present
+        // We strip leading/trailing comments of each token present, but keep whitespace
+        let string = item
+            .tokens()
+            .map(|token| {
+                token
+                    .update_trivia(
+                        FormatTriviaType::Replace(
+                            token
+                                .leading_trivia()
+                                .filter(|token| !trivia_is_comment(token))
+                                .map(|x| x.to_owned())
+                                .collect(),
+                        ),
+                        FormatTriviaType::Replace(
+                            token
+                                .trailing_trivia()
+                                .filter(|token| !trivia_is_comment(token))
+                                .map(|x| x.to_owned())
+                                .collect(),
+                        ),
+                    )
+                    .to_string()
+            })
+            .collect::<String>();
+
         let lines = string.lines();
 
         lines.enumerate().any(|(idx, line)| {