+++ tests/inputs-luau-full_moon/types_indexable.lua
@@ -2,3 +2,4 @@ local x: module.Foo = nil
 local x: module.Array<string> = { "bar" }
 local x: module.Foo | string = "bar"
 local x: module.Foo? = nil
+local x: module.Foo<...string> = nil